+++ internal/printer/json2table/jsonpaths/kafka.go
@@ -8,8 +8,9 @@ var (
 		"Size":            "properties.size",
 		"DatacenterId":    "properties.connections.0.datacenterId",
 		"LanId":           "properties.connections.0.lanId",
-		"BrokerAddresses": "properties.connections.0.brokers",
+		"BrokerAddresses": "properties.connections.0.brokerAddresses",
 		"State":           "metadata.state",
+		"StateMessage":    "metadata.message",
 	}
 
 	KafkaTopic = map[string]string{
@@ -21,5 +22,10 @@ var (
 		"SegmentByes":        "properties.logRetention.segmentBytes",
 		"State":              "metadata.state",
 		"ClusterId":          "href",
+	}
+
+	KafkaUser = map[string]string{
+		"Id":   "id",
+		"Name": "properties.name",
 	}
 )